Fix race condition in beforeSpec
